+++ Components/Misc/VehicleSeat.lua
@@ -25,8 +25,6 @@
 --]]
 local _, ns = ...
 
-if (not ns.IsRetail and not ns.IsCata) then return end
-
 local L = LibStub("AceLocale-3.0"):GetLocale((...))
 
 local VehicleSeat = ns:NewModule("VehicleSeat", ns.MovableModulePrototype, "LibMoreEvents-1.0", "AceHook-3.0")