+++ Components/UnitFrames/Units/NamePlates.lua
@@ -2978,7 +2978,7 @@ cvars = {
 
 	-- The max distance to show nameplates.
 	-- *this value can be set by the user, and all other values are relative to this one.
-	["nameplateMaxDistance"] = ns.IsRetail and NAMEPLATE_MAX_DISTANCE_DEFAULT or ns.IsClassic and 20 or 40, -- Wrath and TBC have 41
+	["nameplateMaxDistance"] = NAMEPLATE_MAX_DISTANCE_DEFAULT,
 
 	-- The maximum distance from the camera (not char) where plates will still have max scale
 	["nameplateMaxScaleDistance"] = 10,