+++ Core/ExplorerMode.lua
@@ -631,13 +631,11 @@ ExplorerMode.OnEvent = function(self, event, ...)
 		self:CheckGroup()
 		self:CheckTarget()
 
-		if (ns.IsRetail or ns.IsCata) then
-			self:CheckFocus()
-			self:CheckVehicle()
-			self:CheckOverride()
-			self:CheckPossess()
-			self:CheckDragonRiding()
-		end
+		self:CheckFocus()
+		self:CheckVehicle()
+		self:CheckOverride()
+		self:CheckPossess()
+		self:CheckDragonRiding()
 
 		self:CheckHealth()
 		self:CheckPower()
@@ -725,17 +723,15 @@ ExplorerMode.EnableExplorerMode = function(self)
 	self:RegisterUnitEvent("UNIT_DISPLAYPOWER", "OnEvent", "player")
 	self:RegisterUnitEvent("UNIT_AURA", "OnEvent", "player", "vehicle")
 
-	if (ns.IsRetail or ns.IsCata) then
-		self:RegisterEvent("PLAYER_FOCUS_CHANGED", "OnEvent")
-		self:RegisterEvent("UPDATE_OVERRIDE_ACTIONBAR", "OnEvent")
-		self:RegisterEvent("UPDATE_POSSESS_BAR", "OnEvent")
-		self:RegisterEvent("UPDATE_BONUS_ACTIONBAR", "OnEvent")
-		self:RegisterEvent("UPDATE_VEHICLE_ACTIONBAR", "OnEvent", "player")
-		self:RegisterUnitEvent("UNIT_ENTERED_VEHICLE", "OnEvent", "player")
-		self:RegisterUnitEvent("UNIT_ENTERING_VEHICLE", "OnEvent", "player")
-		self:RegisterUnitEvent("UNIT_EXITED_VEHICLE", "OnEvent", "player")
-		self:RegisterUnitEvent("UNIT_EXITING_VEHICLE", "OnEvent", "player")
-	end
+	self:RegisterEvent("PLAYER_FOCUS_CHANGED", "OnEvent")
+	self:RegisterEvent("UPDATE_OVERRIDE_ACTIONBAR", "OnEvent")
+	self:RegisterEvent("UPDATE_POSSESS_BAR", "OnEvent")
+	self:RegisterEvent("UPDATE_BONUS_ACTIONBAR", "OnEvent")
+	self:RegisterEvent("UPDATE_VEHICLE_ACTIONBAR", "OnEvent", "player")
+	self:RegisterUnitEvent("UNIT_ENTERED_VEHICLE", "OnEvent", "player")
+	self:RegisterUnitEvent("UNIT_ENTERING_VEHICLE", "OnEvent", "player")
+	self:RegisterUnitEvent("UNIT_EXITED_VEHICLE", "OnEvent", "player")
+	self:RegisterUnitEvent("UNIT_EXITING_VEHICLE", "OnEvent", "player")
 
 	-- This is needed to put actionbars that were exempt from Explorer Mode fading
 	-- back into it when their own full fading has been disabled.
@@ -759,17 +755,15 @@ ExplorerMode.DisableExplorerMode = function(self)
 	self:UnregisterEvent("UNIT_DISPLAYPOWER", "OnEvent", "player")
 	self:UnregisterEvent("UNIT_AURA", "OnEvent", "player", "vehicle")
 
-	if (ns.IsRetail or ns.IsCata) then
-		self:UnregisterEvent("PLAYER_FOCUS_CHANGED", "OnEvent")
-		self:UnregisterEvent("UPDATE_OVERRIDE_ACTIONBAR", "OnEvent")
-		self:UnregisterEvent("UPDATE_POSSESS_BAR", "OnEvent")
-		self:UnregisterEvent("UPDATE_BONUS_ACTIONBAR", "OnEvent")
-		self:UnregisterEvent("UPDATE_VEHICLE_ACTIONBAR", "OnEvent", "player")
-		self:UnregisterEvent("UNIT_ENTERED_VEHICLE", "OnEvent", "player")
-		self:UnregisterEvent("UNIT_ENTERING_VEHICLE", "OnEvent", "player")
-		self:UnregisterEvent("UNIT_EXITED_VEHICLE", "OnEvent", "player")
-		self:UnregisterEvent("UNIT_EXITING_VEHICLE", "OnEvent", "player")
-	end
+	self:UnregisterEvent("PLAYER_FOCUS_CHANGED", "OnEvent")
+	self:UnregisterEvent("UPDATE_OVERRIDE_ACTIONBAR", "OnEvent")
+	self:UnregisterEvent("UPDATE_POSSESS_BAR", "OnEvent")
+	self:UnregisterEvent("UPDATE_BONUS_ACTIONBAR", "OnEvent")
+	self:UnregisterEvent("UPDATE_VEHICLE_ACTIONBAR", "OnEvent", "player")
+	self:UnregisterEvent("UNIT_ENTERED_VEHICLE", "OnEvent", "player")
+	self:UnregisterEvent("UNIT_ENTERING_VEHICLE", "OnEvent", "player")
+	self:UnregisterEvent("UNIT_EXITED_VEHICLE", "OnEvent", "player")
+	self:UnregisterEvent("UNIT_EXITING_VEHICLE", "OnEvent", "player")
 
 	-- This is needed to put actionbars that were exempt from Explorer Mode fading
 	-- back into it when their own full fading has been disabled.